requirements.txt
quart
hypercorn
a2s
mcstatus
pillow